Make GCC happy with printf formats

+++ src/client_side.cc
@@ -1,6 +1,6 @@
 
 /*
- * $Id: client_side.cc,v 1.575 2002/04/14 21:51:36 hno Exp $
+ * $Id: client_side.cc,v 1.576 2002/04/14 22:08:36 hno Exp $
  *
  * DEBUG: section 33    Client-side Routines
  * AUTHOR: Duane Wessels
@@ -1762,7 +1762,7 @@ clientWriteComplete(int fd, char *bufnotused, size_t size, int errflag, void *da
     if (http->out.size > 0x7FFF0000) {
 	debug(33, 1) ("WARNING: closing FD %d to prevent counter overflow\n", fd);
 	debug(33, 1) ("\tclient %s\n", inet_ntoa(http->conn->peer.sin_addr));
-	debug(33, 1) ("\treceived %d bytes\n", http->out.size);
+	debug(33, 1) ("\treceived %d bytes\n", (int)http->out.size);
 	debug(33, 1) ("\tURI %s\n", http->log_uri);
 	comm_close(fd);
     } else
@@ -1771,8 +1771,8 @@ clientWriteComplete(int fd, char *bufnotused, size_t size, int errflag, void *da
     if (http->out.offset > 0x7FFF0000) {
 	debug(33, 1) ("WARNING: closing FD %d to prevent counter overflow\n", fd);
 	debug(33, 1) ("\tclient %s\n", inet_ntoa(http->conn->peer.sin_addr));
-	debug(33, 1) ("\treceived %d bytes (offset %d)\n", http->out.size,
-	    http->out.offset);
+	debug(33, 1) ("\treceived %d bytes (offset %d)\n", (int)http->out.size,
+	    (int)http->out.offset);
 	debug(33, 1) ("\tURI %s\n", http->log_uri);
 	comm_close(fd);
     } else
